+++ android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.debindradhakal199.nepali_audio_library">
    <application
-        android:label="Nepali audio library"
+        android:label="nepali_audio_library"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
