+++ lib/load_audio_book_screen.dart
@@ -52,11 +52,10 @@ class LoadAudioBookScreen extends StatelessWidget {
             color: Colors.black87,
           ),
           child: Card(
-            color: Colors.black,
             child: ListTile(
               title: Text(
                 state.audioBooks[index].bookTitle,
-                style: TextStyle(fontSize: 28, fontWeight: FontWeight.bold, color: Colors.white),
+                style: TextStyle(fontSize: 28, fontWeight: FontWeight.bold),
               ),
               subtitle: Text(
                 state.audioBooks[index].bookLabel,
