+++ macos/Runner/Configs/AppInfo.xcconfig
@@ -5,7 +5,7 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = Nepali audio library;
+PRODUCT_NAME = nepali_audio_library
 
 // The application's bundle identifier
 PRODUCT_BUNDLE_IDENTIFIER = com.debindradhakal199.nepaliAudioLibrary
