+++ web/index.html
@@ -29,7 +29,7 @@
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>Nepali audio library</title>
+  <title>nepali_audio_library</title>
   <link rel="manifest" href="manifest.json">
 
   <script>
